fix: issue with parsing empty string
