+++ src/components/main/Main.css
@@ -28,7 +28,7 @@
   background-position: center;
   background-size: 50%;
   background-repeat: no-repeat;
-  background-image: url('../assets/search-icon.svg');
+  background-image: url('../../assets/search-icon.svg');
 }
 
 .light-block {
